+++ src/parser/part/prelude.rs
@@ -5,8 +5,8 @@ pub(crate) use crate::ast::{
     Expression, InOut, InterfaceEnd, InterfaceEndMember, InterfaceEndReferenceOperator,
     InterfaceEndTarget, InterfacePart, InterfaceUsage, InterfaceUsageBodyElement, Membership, Node,
     PartDef, PartDefBody, PartDefBodyElement, PartUsage, PartUsageBody, PartUsageBodyElement,
-    Perform, PerformBody, PerformBodyElement, PerformInOutBinding, RefBody, RefDecl,
-    VariantTypedUsage, VariantUsage, VariantUsageForm,
+    Perform, PerformActionTarget, PerformBody, PerformBodyElement, PerformInOutBinding, RefBody,
+    RefDecl, VariantTypedUsage, VariantUsage, VariantUsageForm,
 };
 pub(crate) use crate::parser::action::{action_def, action_usage};
 pub(crate) use crate::parser::allocation::{allocation_def, allocation_usage};